+++ demo/meeting/src/containers/MeetingRoster.tsx
@@ -2,23 +2,15 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React, { useState, ChangeEvent } from 'react';
-import styled from 'styled-components';
 import {
   Roster,
   RosterHeader,
   RosterGroup,
+  Cell,
   useRoster
 } from 'amazon-chime-sdk-component-library-react';
 
 import RosterAttendee from './RosterAttendee';
-
-const StyledRoster = styled.div`
-  position: fixed;
-  right: 0;
-  top: 0;
-  bottom: 0;
-  width: 18.5rem;
-`;
 
 const MeetingRoster = () => {
   const roster = useRoster();
@@ -42,7 +34,7 @@ const MeetingRoster = () => {
   });
 
   return (
-    <StyledRoster>
+    <Cell gridArea="roster">
       <Roster>
         <RosterHeader
           searchValue={filter}
@@ -53,7 +45,7 @@ const MeetingRoster = () => {
         />
         <RosterGroup>{attendeeItems}</RosterGroup>
       </Roster>
-    </StyledRoster>
+    </Cell>
   );
 };