+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Import components relatively using alias
 - Update demo app README
 - Update `Modal` to SDK Modal component and minor change to `Card` component in the demo
+- Updating demo with basic video grid
 
 ### Removed
 - Removed active state button tests.